code/host/requirements.txt
appjar
pyserial
